Relocate batch mutations graphql tests
